requirements.txt
loguru
influxdb
python-kasa
python-logging-loki